# ent/schema/application.go
// Code generated by cmd/entgen from CUE ontology. DO NOT EDIT.
package schema

import (
	"regexp"

	"entgo.io/ent"
	"entgo.io/ent/schema/edge"
	"entgo.io/ent/schema/field"
	"entgo.io/ent/schema/mixin"
	"github.com/google/uuid"
)

// Ensure mixin import is used.
var _ mixin.Schema

// Application holds the schema definition for the Application entity.
type Application struct {
	ent.Schema
}

// Mixin of the Application.
func (Application) Mixin() []ent.Mixin {
	return []ent.Mixin{
		AuditMixin{},
	}
}

// Fields of the Application.
func (Application) Fields() []ent.Field {
	return []ent.Field{
		field.UUID("id", uuid.UUID{}).Default(uuid.New).Immutable().Comment("Primary key"),
		field.UUID("applicant_person_id", uuid.UUID{}),
		field.Enum("status").Values("submitted", "screening", "under_review", "approved", "conditionally_approved", "denied", "withdrawn", "expired"),
		field.Time("desired_move_in"),
		field.Int("desired_lease_term_months"),
		field.String("screening_request_id").Optional().Nillable().SchemaType(map[string]string{"postgres": "varchar"}),
		field.Time("screening_completed").Optional().Nillable(),
		field.Int("credit_score").Optional().Nillable(),
		field.Bool("background_clear").Default(false),
		field.Bool("income_verified").Default(false),
		field.Float("income_to_rent_ratio").Optional().Nillable(),
		field.String("decision_by").Optional().Nillable().SchemaType(map[string]string{"postgres": "varchar"}),
		field.Time("decision_at").Optional().Nillable(),
		field.String("decision_reason").Optional().Nillable().SchemaType(map[string]string{"postgres": "varchar"}),
		field.JSON("conditions", []string{}).Optional(),
		field.Int64("application_fee_amount_cents").Comment("application_fee — amount in cents"),
		field.String("application_fee_currency").Default("USD").Match(regexp.MustCompile(`^[A-Z]{3}$`)).Comment("application_fee — ISO 4217 currency code"),
		field.Bool("fee_paid").Default(false),
	}
}

// Edges of the Application.
func (Application) Edges() []ent.Edge {
	return []ent.Edge{
		edge.From("property", Property.Type).Ref("applications").Unique().Comment("Property receives Applications (inverse)"),
		edge.From("space", Space.Type).Ref("applications").Unique().Comment("Space receives Applications (inverse)"),
		edge.From("resulting_lease", Lease.Type).Ref("application").Unique().Comment("Lease originated from Application (inverse)"),
		edge.To("applicant", Person.Type).Unique().Required().Field("applicant_person_id").Comment("Application was submitted by Person"),
	}
}

// ValidApplicationTransitions defines the allowed state machine transitions.
// Generated from CUE ontology state_machines.cue.
var ValidApplicationTransitions = map[string][]string{
	"approved":               {"expired"},
	"conditionally_approved": {"approved", "denied", "withdrawn", "expired"},
	"denied":                 {},
	"expired":                {},
	"screening":              {"under_review", "withdrawn"},
	"submitted":              {"screening", "withdrawn"},
	"under_review":           {"approved", "conditionally_approved", "denied", "withdrawn"},
	"withdrawn":              {},
}