# gen/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "description": "REST API generated from CUE ontology. All endpoints accept/return JSON.",
        "title": "Propeller Property Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "description": "Local development",
            "url": "http://localhost:8080"
        }
    ],
    "paths": {
...
        "/v1/applications/{id}/approve": {
            "post": {
                "operationId": "ApproveApplication",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "409": {
                        "description": "Invalid State Transition"
                    }
                },
                "summary": "Approve a lease application (transitions to \"approved\")",
                "tags": [
                    "Application"
                ]
            }
        },
        "/v1/applications/{id}/deny": {
            "post": {
                "operationId": "DenyApplication",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Application"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "409": {
                        "description": "Invalid State Transition"
                    }
                },
                "summary": "Deny a lease application (transitions to \"denied\")",
                "tags": [
                    "Application"
                ]
            }
        },
...
        "/v1/journal-entries/{id}/post": {
            "post": {
                "operationId": "PostJournalEntry",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JournalEntry"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "409": {
                        "description": "Invalid State Transition"
                    }
                },
                "summary": "Post a journal entry. Lines must balance (debits = credits) (transitions to \"posted\")",
                "tags": [
                    "JournalEntry"
                ]
            }
        },
        "/v1/journal-entries/{id}/void": {
            "post": {
                "operationId": "VoidJournalEntry",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JournalEntry"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "409": {
                        "description": "Invalid State Transition"
                    }
                },
                "summary": "Void a posted journal entry. Creates reversal entry (transitions to \"voided\")",
                "tags": [
                    "JournalEntry"
                ]
            }
        },
...
        "/v1/reconciliations/{id}/approve": {
            "post": {
                "operationId": "ApproveReconciliation",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reconciliation"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "409": {
                        "description": "Invalid State Transition"
                    }
                },
                "summary": "Approve a balanced reconciliation (transitions to \"approved\")",
                "tags": [
                    "Reconciliation"
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Application": {
                "type": "object",
                "properties": {
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "property_id": {
                        "type": "string"
                    },
                    "unit_id": {
                        "type": "string"
                    },
                    "applicant_person_id": {
                        "type": "string"
                    },
                    "status": {
                        "enum": [
                            "submitted",
                            "screening",
                            "under_review",
                            "approved",
                            "conditionally_approved",
                            "denied",
                            "withdrawn",
                            "expired"
                        ],
                        "type": "string"
                    },
                    "desired_move_in": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "desired_lease_term_months": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "screening_request_id": {
                        "type": "string"
                    },
                    "screening_completed": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "credit_score": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "background_clear": {
                        "type": "boolean"
                    },
                    "income_verified": {
                        "type": "boolean"
                    },
                    "income_to_rent_ratio": {
                        "format": "double",
                        "type": "number"
                    },
                    "decision_by": {
                        "type": "string"
                    },
                    "decision_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "decision_reason": {
                        "type": "string"
                    },
                    "conditions": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "application_fee_amount_cents": {
                        "description": "application_fee amount in cents",
                        "format": "int64",
                        "type": "integer"
                    },
                    "application_fee_currency": {
                        "description": "application_fee ISO 4217 currency code",
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                    },
                    "fee_paid": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "updated_by": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "property_id",
                    "applicant_person_id",
                    "status",
                    "desired_move_in",
                    "desired_lease_term_months",
                    "background_clear",
                    "income_verified",
                    "application_fee_amount_cents",
                    "fee_paid"
                ]
            },
...
            "JournalEntry": {
                "type": "object",
                "properties": {
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "entry_date": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "posted_date": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "source_type": {
                        "enum": [
                            "manual",
                            "auto_charge",
                            "payment",
                            "bank_import",
                            "cam_reconciliation",
                            "depreciation",
                            "accrual",
                            "intercompany",
                            "management_fee",
                            "system"
                        ],
                        "type": "string"
                    },
                    "source_id": {
                        "type": "string"
                    },
                    "status": {
                        "enum": [
                            "draft",
                            "pending_approval",
                            "posted",
                            "voided"
                        ],
                        "type": "string"
                    },
                    "approved_by": {
                        "type": "string"
                    },
                    "approved_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "batch_id": {
                        "type": "string"
                    },
                    "entity_id": {
                        "type": "string"
                    },
                    "property_id": {
                        "type": "string"
                    },
                    "reverses_journal_id": {
                        "type": "string"
                    },
                    "reversed_by_journal_id": {
                        "type": "string"
                    },
                    "lines": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array"
                    },
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "updated_by": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "entry_date",
                    "posted_date",
                    "description",
                    "source_type",
                    "status",
                    "lines"
                ]
            },
...
            "Reconciliation": {
                "type": "object",
                "properties": {
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "bank_account_id": {
                        "type": "string"
                    },
                    "period_start": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "period_end": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "statement_balance_amount_cents": {
                        "description": "statement_balance amount in cents",
                        "format": "int64",
                        "type": "integer"
                    },
                    "statement_balance_currency": {
                        "description": "statement_balance ISO 4217 currency code",
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                    },
                    "system_balance_amount_cents": {
                        "description": "system_balance amount in cents",
                        "format": "int64",
                        "type": "integer"
                    },
                    "system_balance_currency": {
                        "description": "system_balance ISO 4217 currency code",
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                    },
                    "difference_amount_cents": {
                        "description": "difference amount in cents",
                        "format": "int64",
                        "type": "integer"
                    },
                    "difference_currency": {
                        "description": "difference ISO 4217 currency code",
                        "pattern": "^[A-Z]{3}$",
                        "type": "string"
                    },
                    "status": {
                        "enum": [
                            "in_progress",
                            "balanced",
                            "unbalanced",
                            "approved"
                        ],
                        "type": "string"
                    },
                    "matched_transaction_count": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "unmatched_transaction_count": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "completed_by": {
                        "type": "string"
                    },
                    "completed_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "approved_by": {
                        "type": "string"
                    },
                    "approved_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "updated_by": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "bank_account_id",
                    "period_start",
                    "period_end",
                    "statement_balance_amount_cents",
                    "system_balance_amount_cents",
                    "difference_amount_cents",
                    "status",
                    "matched_transaction_count",
                    "unmatched_transaction_count"
                ]
            },
...
        }
    }
}